+++ src/main/java/hungteen/opentd/common/impl/requirement/NotRequirement.java
@@ -5,8 +5,6 @@ import com.mojang.serialization.codecs.RecordCodecBuilder;
 import hungteen.htlib.util.helper.PlayerHelper;
 import hungteen.opentd.api.interfaces.ISummonRequirement;
 import hungteen.opentd.api.interfaces.ISummonRequirementType;
-import hungteen.opentd.api.interfaces.ITargetFilter;
-import hungteen.opentd.common.impl.filter.HTTargetFilters;
 import net.minecraft.core.BlockPos;
 import net.minecraft.network.chat.Component;
 import net.minecraft.server.level.ServerLevel;
@@ -14,24 +12,24 @@ import net.minecraft.world.entity.Entity;
 import net.minecraft.world.entity.player.Player;
 import net.minecraft.world.level.block.state.BlockState;
 
-import java.util.Objects;
+import java.util.List;
 import java.util.Optional;
 
 /**
  * @program: HTOpenTD
  * @author: HungTeen
- * @create: 2023-01-07 20:30
+ * @create: 2023-01-17 11:22
  **/
-public record EntityRequirement(Optional<String> tip, ITargetFilter filter) implements ISummonRequirement {
+public record NotRequirement(ISummonRequirement requirement, Optional<String> tip) implements ISummonRequirement {
 
-    public static final Codec<EntityRequirement> CODEC = RecordCodecBuilder.<EntityRequirement>mapCodec(instance -> instance.group(
-            Codec.optionalField("tip", Codec.STRING).forGetter(EntityRequirement::tip),
-            HTTargetFilters.getCodec().fieldOf("filter").forGetter(EntityRequirement::filter)
-            ).apply(instance, EntityRequirement::new)).codec();
+    public static final Codec<NotRequirement> CODEC = RecordCodecBuilder.<NotRequirement>mapCodec(instance -> instance.group(
+            HTSummonRequirements.getCodec().fieldOf("requirement").forGetter(NotRequirement::requirement),
+            Codec.optionalField("tip", Codec.STRING).forGetter(NotRequirement::tip)
+    ).apply(instance, NotRequirement::new)).codec();
 
     @Override
     public boolean allowOn(ServerLevel level, Player player, Entity entity, boolean sendMessage) {
-        if(! filter().match(level, player, entity)){
+        if(requirement().allowOn(level, player, entity, false)){
             if(sendMessage) PlayerHelper.sendTipTo(player, getTip());
             return false;
         }
@@ -40,7 +38,11 @@ public record EntityRequirement(Optional<String> tip, ITargetFilter filter) impl
 
     @Override
     public boolean allowOn(ServerLevel level, Player player, BlockState state, BlockPos pos, boolean sendMessage) {
-        return false;
+        if(requirement().allowOn(level, player, state, pos, false)){
+            if(sendMessage) PlayerHelper.sendTipTo(player, getTip());
+            return false;
+        }
+        return true;
     }
 
     @Override
@@ -48,11 +50,11 @@ public record EntityRequirement(Optional<String> tip, ITargetFilter filter) impl
     }
 
     public Component getTip() {
-        return this.tip().map(Component::translatable).orElse(Component.translatable("tip.opentd.wrong_entity"));
+        return this.tip().map(Component::translatable).orElse(Component.empty());
     }
 
     @Override
     public ISummonRequirementType<?> getType() {
-        return HTSummonRequirements.ENTITY_REQUIREMENT;
+        return HTSummonRequirements.NOT_REQUIREMENT;
     }
 }